backend/pyproject.toml
[tool.poetry]
name = "pos-backend"
version = "0.1.0"
description = ""
authors = ["Vincent Schmid <Vincent.Schmid@ti8m.ch>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
gunicorn = "^20.1.0"
django-ninja = "^0.19.1"
django-ninja-extra = "^0.15.4"
django-ninja-jwt = "^5.2.0"
Markdown = "^3.4.1"
django-filter = "^22.1"
django-cors-headers = "^3.13.0"
dj-database-url = "^1.0.0"
psycopg2 = "^2.9.3"
psycopg = {extras = ["binary"], version = "^3.0.16"}
django-storages = {extras = ["azure"], version = "^1.13.1"}
django-environ = "^0.9.0"
eventsourcing = "^9.2.16"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
coverage = "^6.4.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
